separated track generation into different files
